Added Atom command and keymap for easier use
Instead of keymaps/markclip.cson, you could also allow the grammars to be set by the user in the config and then check to see if the editor grammar is in this set before handling an insert event. This is how `markdown-preview-plus` works. The (possible) downside is that the command is activated for every cmd/ctrl-z event which some people may not like.